requirements.txt
numpy
torch
torchvision
pillow
opencv-python  # or opencv-headless-python / opencv-contrib-python
albumentations
pyyaml

matplotlib
jupyter
tqdm

onnx
onnxruntime
onnxscript

scikit-learn
timm
yacs
termcolor
wandb
seaborn
tensorboard